feat：实现/api/rooms

+++ src/main/java/com/example/DisneyBookingBackend/models/dto/AvailableThemeRoomDto.java
@@ -4,13 +4,13 @@ package com.example.DisneyBookingBackend.models.dto;
 public class AvailableThemeRoomDto {
     private String themeName;
     private Integer availableRoomCount;
-    private Double averageRating;
+    private Float averageRating;
     private SampleRoomDto sampleRoom;
 
     public AvailableThemeRoomDto() {
     }
 
-    public AvailableThemeRoomDto(String themeName, Integer availableRoomCount, Double averageRating, SampleRoomDto sampleRoom) {
+    public AvailableThemeRoomDto(String themeName, Integer availableRoomCount, Float averageRating, SampleRoomDto sampleRoom) {
         this.themeName = themeName;
         this.availableRoomCount = availableRoomCount;
         this.averageRating = averageRating;
@@ -33,11 +33,11 @@ public class AvailableThemeRoomDto {
         this.availableRoomCount = availableRoomCount;
     }
 
-    public Double getAverageRating() {
+    public Float getAverageRating() {
         return averageRating;
     }
 
-    public void setAverageRating(Double averageRating) {
+    public void setAverageRating(Float averageRating) {
         this.averageRating = averageRating;
     }
